[refactor]: separate Express app and Socket.IO server into distinct files
- chore: create dedicated ioApp.js for Socket.IO functionality
- update: modify server.js to use HTTP server with express app and socket.io server
